requirements.txt
starlette==0.20.4
scikit-surprise==1.1.3
pandas==2.0.0
numpy==1.23.5